feat: show cutlines and refactor
Vertical lines corresponding to the cut coordinates are shown if desired
in each of the Axes object.
The drawing settings of these lines is handled in updateCoordCut()
method.

In addition the redraw() method now only updates the display and is
called whenever needed (with major performance improvements).
The code that updates the curves is placed in the updateCurves() method.